Quote GOROOT wasm_exec path to avoid splitting on spaces
My GOROOT path doesn't have any spaces in it, but somebody else's might
do and then this path would be split at the spaces and used by the `cp`
command as separate arguments, of which likely none of them would be
valid paths.  Quoting it treats it as a single argument with spaces.